Made initial version of brick's spawn
Added BrickSet class which will manage everything related to bricks. For now, it construct a matrix occupying the whole width of the screen and maximum the paddle's Y position minus 4 times the paddle's height in vertical. Still need to apply collision with the paddle and management to specific positions of the matrix