Implemented basic collision system between ball and the bricks
Each frame, each ball checks all the bricks to see if it collided with any of them. If so, that brick is set to 0 in the matrix and is not checked any further. Also improved performance by adding another rule to Ball when checking collision with the paddle: it only does it when the ball's Y velocity is positive (the ball is going down the screen)